+++ lib/core/impl/CoreServiceImpl.ts
@@ -8,7 +8,7 @@ import { MetadataServiceImpl } from "@/lib/core/impl/MetadataServiceImpl"
 import { SynchronizedContextService } from "@/lib/core/impl/SynchronizedContextService"
 import { isDataEntity } from "@/lib/utils"
 import { DateTime } from "luxon"
-import { IProfileService } from "@/lib/core/profiles/IProfileService"
+import { IProfileService, ProfileServiceOptions } from "@/lib/core/profiles/IProfileService"
 import { ProfileService } from "@/lib/core/profiles/impl/ProfileService"
 
 /**
@@ -144,11 +144,17 @@ export class CoreServiceImpl implements ICoreService {
 
     static async newInstance(
         persistenceAdapter: IPersistenceAdapter,
-        crateService: ICrateService
+        crateService: ICrateService,
+        profileServiceOptions: ProfileServiceOptions = {}
     ): Promise<CoreServiceImpl> {
         const metadata = await MetadataServiceImpl.newInstance(persistenceAdapter)
         const context =
             await SynchronizedContextService.newInstanceWithPersistence(persistenceAdapter)
-        return new CoreServiceImpl(metadata, context, new ProfileService(metadata), crateService)
+        return new CoreServiceImpl(
+            metadata,
+            context,
+            new ProfileService(metadata, profileServiceOptions),
+            crateService
+        )
     }
 }