+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added profile handling to NovaCrate
 - NovaCrate will attempt to load any referenced profiles via the `conformsTo` property of the root entity
-  - Note that this is only expected to work for the Workflow RO-Crate Profile at the moment
+  - User approval must be given for unknown profile URLs, which can be done through a new section in the settings
 - Active profiles enhance the editor in many different places, such as the modal for creating entities and the modal for adding properties to entities
 - Added a profiles section to the context page to understand active profiles and their effects
 - Added validation for profiles based on the RO-Crate MASP proposal